Patching parser with safe functionality
Names are less rigid so they don't have to be namespaced (this was mainly due to computronics). If they are not it will ask you to specify the owner which is a bit better. I also found out that Ctrl+C doesn't work properly with prompts so added a quit thing but it currently only works on getting the name. Need to make it more generic later on but it gives a bail out point that is always shown before writing the file